eighth commit, valid tests
