+++ demo/main.cpp
@@ -1,73 +1,73 @@
 #include <student.hpp>
 
-void print(const Student& student, std::ostream& os){
-    if (student.name.empty()){
-        os << "| null\t|";
-    } else {
-        os << "| " << student.name << "\t|";
-    }
-
-    if (student.group.type() == typeid(std::nullptr_t)){
-        os << "null|";
-    } else if (student.group.type() == typeid(std::string)){
-        if (std::any_cast<std::string>(student.group).length() == 1){
-            os << " " << std::any_cast<std::string>(student.group) << "\t\t\t|";
-        } else {
-            os << " " << std::any_cast<std::string>(student.group) << "\t|";
-        }
-    } else {
-        os << " " << std::any_cast<std::size_t>(student.group)
-           << " group\t|";
-    }
-
-    if (student.avg.type() == typeid(std::nullptr_t)){
-        os << "null\t|";
-    } else if (student.avg.type() == typeid(std::string)){
-        os << " " << std::any_cast<std::string>(student.avg) << "\t|";
-    } else if (student.avg.type() == typeid(std::double_t)){
-        os << " " << std::any_cast<std::double_t>(student.avg) << "\t|";
-    } else {
-        os << " " << std::any_cast<std::size_t>(student.avg) << "\t\t|";
-    }
-
-    if (student.debt.type() == typeid(std::nullptr_t)) {
-        os << " null\t\t|" << std::endl;
-    } else if (student.debt.type() == typeid(std::string)) {
-        os << " " << std::any_cast<std::string>(student.debt) << "\t\t|" <<  std::endl;
-    } else {
-        os
-                << " " << std::any_cast<std::vector<std::string> >(student.debt).size()
-                << " items\t|" << std::endl;
-    }
-}
-
-void print(const std::vector<Student>& students, std::ostream& os) {
-    std::cout << "| name          | group     | avg   | debt      |" << std::endl;
-    std::cout << "|---------------|-----------|-------|-----------|" << std::endl;
-    for (auto const& student : students) {
-        print(student, os);
-        std::cout << "|---------------|-----------|-------|-----------|" << std::endl;
-    }
-}
-
-std::vector<Student> read_file(json data) {
-    std::vector<Student> students;
-    for (auto const &item : data.at("items")) {
-        Student student;
-        from_json(item, student);
-        students.push_back(student);
-    }
-    return students;
-}
-
-int main() {
-    std::string jsonPath;
-    jsonPath = "/Users/pvelp/students.json";
-//    if (argc < 1){
-//        std::cout << "Не передали путь к файлу" << std::endl;
+//void print(const Student& student, std::ostream& os){
+//    if (student.name.empty()){
+//        os << "| null\t|";
 //    } else {
-//        jsonPath = argv[1];
+//        os << "| " << student.name << "\t|";
 //    }
+//
+//    if (student.group.type() == typeid(std::nullptr_t)){
+//        os << "null|";
+//    } else if (student.group.type() == typeid(std::string)){
+//        if (std::any_cast<std::string>(student.group).length() == 1){
+//            os << " " << std::any_cast<std::string>(student.group) << "\t\t\t|";
+//        } else {
+//            os << " " << std::any_cast<std::string>(student.group) << "\t|";
+//        }
+//    } else {
+//        os << " " << std::any_cast<std::size_t>(student.group)
+//           << " group\t|";
+//    }
+//
+//    if (student.avg.type() == typeid(std::nullptr_t)){
+//        os << "null\t|";
+//    } else if (student.avg.type() == typeid(std::string)){
+//        os << " " << std::any_cast<std::string>(student.avg) << "\t|";
+//    } else if (student.avg.type() == typeid(std::double_t)){
+//        os << " " << std::any_cast<std::double_t>(student.avg) << "\t|";
+//    } else {
+//        os << " " << std::any_cast<std::size_t>(student.avg) << "\t\t|";
+//    }
+//
+//    if (student.debt.type() == typeid(std::nullptr_t)) {
+//        os << " null\t\t|" << std::endl;
+//    } else if (student.debt.type() == typeid(std::string)) {
+//        os << " " << std::any_cast<std::string>(student.debt) << "\t\t|" <<  std::endl;
+//    } else {
+//        os
+//                << " " << std::any_cast<std::vector<std::string> >(student.debt).size()
+//                << " items\t|" << std::endl;
+//    }
+//}
+//
+//void print(const std::vector<Student>& students, std::ostream& os) {
+//    os << "| name          | group     | avg   | debt      |\n";
+//    os << "|---------------|-----------|-------|-----------|\n";
+//    for (auto const& student : students) {
+//        print(student, os);
+//        os << "|---------------|-----------|-------|-----------|\n";
+//    }
+//}
+//
+//std::vector<Student> read_file(json data) {
+//    std::vector<Student> students;
+//    for (auto const &item : data.at("items")) {
+//        Student student;
+//        from_json(item, student);
+//        students.push_back(student);
+//    }
+//    return students;
+//}
+
+int main(int argc, char* argv[]) {
+    std::string jsonPath;
+//    jsonPath = "/Users/pvelp/students.json";
+    if (argc < 2){
+        throw std::runtime_error("Path is not exist");
+    } else {
+        jsonPath = argv[1];
+    }
 
 
     std::ifstream file{jsonPath};
@@ -78,6 +78,16 @@ int main() {
     json data;
     file >> data;
 
+    if (data["_meta"]["count"] != data["items"].size()){
+        throw std::runtime_error{"_meta doesn't equal number of students"};
+    }
+
+    if (!data["items"].is_array()){
+        throw std::runtime_error{"items is not array"};
+    }
+
+
     std::vector<Student> students = read_file(data);
     print(students, std::cout);
+
 }
