+++ include/student.hpp
@@ -8,6 +8,10 @@
 #include <string>
 #include <fstream>
 #include <vector>
+#include <string>
+#include <sstream>
+#include <iostream>
+#include <gtest/gtest.h>
 
 using nlohmann::json;
 
@@ -29,5 +33,7 @@ auto get_group(const json& j) -> std::any;
 void from_json(const json& j, Student& s);
 
 std::vector<Student> read_file(json data);
+void print(const std::vector<Student>& students, std::ostream& os);
+void print(const Student& student, std::ostream& os);
 
 #endif // INCLUDE_STUDENT_HPP_
